integration_test: add tests for compileVariableTTF and CFF2
with and without useProductionNames set
